Rename file and add licence text
